+++ examples/example_image.py
@@ -1,16 +1,10 @@
 #!/usr/bin/python3
-# -*- coding: utf-8 -*-
 import pprint
+
 import serpscrap
 
-keywords = ['lost places']
-
-config = serpscrap.Config()
-config.set('search_type', 'image')
-
 scrap = serpscrap.SerpScrap()
-scrap.init(config=config.get(), keywords=keywords)
-results = scrap.run()
+results = scrap.search(['lost places'], search_type='image')
 
 for result in results[:10]:
     pprint.pprint(result)