+++ examples/example_simple.py
@@ -1,15 +1,10 @@
 #!/usr/bin/python3
-# -*- coding: utf-8 -*-
 import pprint
+
 import serpscrap
 
-keywords = ['bienen']
-
-config = serpscrap.Config()
-
 scrap = serpscrap.SerpScrap()
-scrap.init(config=config.get(), keywords=keywords)
-results = scrap.run()
+results = scrap.search(['bienen'])
 
 for result in results:
     pprint.pprint(result)