+++ examples/example_url.py
@@ -1,5 +1,4 @@
 #!/usr/bin/python3
-# -*- coding: utf-8 -*-
 import serpscrap
 
 url = 'https://en.wikipedia.org/wiki/Special:Random'