+++ serpscrap/__init__.py
@@ -1,7 +1,20 @@
 #!/usr/bin/python3
 from serpscrap.config import Config
-from serpscrap.csv_writer import CsvWriter
+from serpscrap.exceptions import ConfigurationError, SerpScrapError
+from serpscrap.models import FailureRecord, SearchReport, SearchRequest
+from serpscrap.output import JsonOutputError, JsonResultWriter
 from serpscrap.serpscrap import SerpScrap
 from serpscrap.urlscrape import UrlScrape
 
-__all__ = ["Config", "CsvWriter", "SerpScrap", "UrlScrape"]
+__all__ = [
+    "Config",
+    "ConfigurationError",
+    "FailureRecord",
+    "JsonOutputError",
+    "JsonResultWriter",
+    "SearchReport",
+    "SearchRequest",
+    "SerpScrap",
+    "SerpScrapError",
+    "UrlScrape",
+]